requirements.txt
numpy
scipy
librosa
soundfile
tensorflow
sounddevice
influxdb_client
paho-mqtt